requirements.txt
keras==2.10.0
matplotlib==3.7.1
numpy==1.24.2
opencv_python==4.7.0.72
Pillow==9.4.0
scikit_learn==1.2.2
segmentation_models==1.0.1
skimage==0.0
tensorflow==2.12.0
tensorflow_intel==2.11.0
